Lot1

+++ src/Lot1/Vehicule.java
@@ -3,10 +3,10 @@ package Lot1;
 public class Vehicule {
 	private String idVehicule;
 	private String modele;
-	private float longueur;
-	private float largeur;
+	private double longueur;
+	private double largeur;
 	
-	public Vehicule(String idVehicule, String modele, float longueur, float largeur) {
+	public Vehicule(String idVehicule, String modele, double longueur, double largeur) {
 		this.idVehicule=idVehicule;
 		this.modele=modele;
 		this.longueur=longueur;
